got delete event to work

+++ Alex/myCalendarProject/app/src/main/java/com/cse110/apk404/myCalendar/eventListHandler/EventListHandler.java
@@ -101,7 +101,7 @@ public class EventListHandler {
     }
 
 
-    public static boolean removeEventById(int temp) throws CalendarError {
+    public static boolean removeEventById(long temp) throws CalendarError {
         boolean check = true;
         if (staticList == null) {
             check = false;
